+++ packages/cms/lib/modules/ideas-on-map-widgets/lib/create-config.js
@@ -19,7 +19,7 @@ module.exports = function createConfig(widget, data, jwt, apiUrl, loginUrl, apos
   // image settings; todo: deze moeten syncen naar de api en dan moet de voorwaardelijkheid omgedraaid
   let allowMultipleImages = typeof widget.imageAllowMultipleImages != 'undefined' ? widget.imageAllowMultipleImages : ( ( data.global.siteConfig && data.global.siteConfig.ideas && typeof data.global.siteConfig.ideas.allowMultipleImages != 'undefined' ) ? data.global.siteConfig.ideas.allowMultipleImages : false );
   let placeholderImageSrc = typeof widget.imagePlaceholderImageSrc != 'undefined' ? apos.attachments.url(widget.imagePlaceholderImageSrc) : ( ( data.global.siteConfig && data.global.siteConfig.ideas && typeof data.global.siteConfig.ideas.placeholderImageSrc != 'undefined' ) ? data.global.siteConfig.ideas.placeholderImageSrc : undefined );
-
+  
   // formfields
   let formFields = [ ...widget.formFields ];
   formFields.forEach((formField) => {
@@ -137,14 +137,14 @@ module.exports = function createConfig(widget, data, jwt, apiUrl, loginUrl, apos
 
     argument: {
       isActive: widget.showReactions,
-      isClosed: data.global.siteConfig && data.global.siteConfig.arguments && typeof data.global.siteConfig.arguments.isClosed != 'undefined' ? data.global.siteConfig.arguments.isClosed : false,
-      closedText: data.global.siteConfig && data.global.siteConfig.arguments && typeof data.global.siteConfig.arguments.closedText != 'undefined' ? data.global.siteConfig.arguments.closedText : true,
       title: widget.reactionsTitle,
       formIntro: widget.reactionsFormIntro,
       placeholder: widget.reactionsPlaceholder,
 			descriptionMinLength: ( data.global.siteConfig && data.global.siteConfig.arguments && data.global.siteConfig.arguments.descriptionMinLength ) || 30,
 			descriptionMaxLength: ( data.global.siteConfig && data.global.siteConfig.arguments && data.global.siteConfig.arguments.descriptionMaxLength ) || 100,
-      closeReactionsForIdeaIds: widget.closeReactionsForIdeaIds,
+      isClosed: !!widget.reactionsClosed,
+      closedText: widget.reactionsClosedText,
+      closeReactionsForIdeaIds: !widget.reactionsClosed && widget.closeReactionsForIdeaIds,
 		},
 
     map: {